requirements.txt
numpy
scipy
opencv-python
python-barcode
pyzbar
qrcode
